+++ guard_agent/__init__.py
@@ -36,7 +36,7 @@ from guard_agent.utils import (
     validate_config,
 )
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"
 
 __all__ = [
     "guard_agent",